requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-cors-headers==3.2.1
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
feedparser==5.2.1
idna==2.9
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
